ingestor/requirements.txt
psycopg2-binary
requests
schedule
komPYoot
